+++ maven/deploy.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-VERSION="2.1.1"
+VERSION="3.0.0"
 
 mvn gpg:sign-and-deploy-file \
 	-DrepositoryId=atvg-studios \