+++ src/main/kotlin/klib/extensions/MutableMap.kt
@@ -5,30 +5,32 @@ package klib.extensions
  *
  * @param other The map to merge
  *
- * @since <NEXT_VERSION>
+ * @since 3.0.0
  * @author Thomas Obernosterer
  */
-fun <K,V> MutableMap<K,V>.smartMerge(other: Map<K,V>) {
+fun <K, V> MutableMap<K, V>.smartMerge(other: Map<K, V>) {
     other.forEach { (key, value) ->
-        if(key !in this) {
+        if (key !in this) {
             this[key] = value
         }
     }
 }
 
 /**
- * Merge a Key and a Value Array into a MutableMap
+ * Merge a Key and a Value Array into a MutableMap; Duplicate keys will be ignored
  *
  * @param keys The Keys
  * @param values The Values
  *
- * @since <NEXT_VERSION>
+ * @since 3.0.0
  * @author Thomas Obernosterer
  */
-fun <K,V> MutableMap<K,V>.mergeArrays(keys: Array<K>, values: Array<V>) {
-    if(keys.size == values.size) {
-        for(i in 0 until keys.size) {
-            this[keys[i]] = values[i]
+fun <K, V> MutableMap<K, V>.mergeArrays(keys: Array<K>, values: Array<V>) {
+    if (keys.size == values.size) {
+        for (i in 0 until keys.size) {
+            if (keys[i] !in this) {
+                this[keys[i]] = values[i]
+            }
         }
     }
 }
@@ -38,10 +40,10 @@ fun <K,V> MutableMap<K,V>.mergeArrays(keys: Array<K>, values: Array<V>) {
  *
  * @param other The map to merge
  *
- * @since <NEXT_VERSION>
+ * @since 3.0.0
  * @author Thomas Obernosterer
  */
-fun <K,V> MutableMap<K,V>.fullMerge(other: Map<K,V>) {
+fun <K, V> MutableMap<K, V>.fullMerge(other: Map<K, V>) {
     other.forEach { (key, value) ->
         this[key] = value
     }