+++ src/main/kotlin/klib/functions/KLibRequire.kt
@@ -7,24 +7,7 @@ import klib.types.SemVer
  * Check if the used kLib is the required version (Dependency Check)
  * If not, kLibInf.exit() is called
  *
- * @param version
- * @see kLibInf.exit
- *
- * @since 0.1.3
- * @author Thomas Obernosterer
- */
-fun kLibRequire(version: Int) {
-    if (kLibInf.versionId != version) {
-        println("This application requires kLib release $version! (Release ${kLibInf.versionId} not Compatible)")
-        kLibInf.exit(1)
-    }
-}
-
-/**
- * Check if the used kLib is the required version (Dependency Check)
- * If not, kLibInf.exit() is called
- *
- * @param version
+ * @param version Required Version
  * @see kLibInf.exit
  *
  * @since 2.1.0
@@ -38,19 +21,19 @@ fun kLibRequire(version: SemVer) {
 }
 
 /**
- * Check if the used kLib is in the required version range (Dependency Check)
+ * Check if the used kLib is the required version (Dependency Check)
  * If not, kLibInf.exit() is called
  *
- * @param minVersion
- * @param maxVersion
+ * @param min Minimum Version
+ * @param max Maximum Version
  * @see kLibInf.exit
  *
- * @since 0.1.3
+ * @since 3.0.0
  * @author Thomas Obernosterer
  */
-fun kLibRequire(minVersion: Int, maxVersion: Int) {
-    if (kLibInf.versionId < minVersion || kLibInf.versionId > maxVersion) {
-        println("This application requires at least kLib release $minVersion and at most kLib release $maxVersion! (Release ${kLibInf.versionId} not Compatible)")
+fun kLibRequire(min: SemVer, max: SemVer) {
+    if (kLibInf.semver < min || kLibInf.semver > max) {
+        println("This application requires kLib a version between $min and $max! (Version ${kLibInf.semver} not Compatible)")
         kLibInf.exit(1)
     }
 }