+++ src/test/kotlin/devtests/FunctionEvery.kt
@@ -1,0 +1,9 @@
+package devtests
+
+import klib.extensions.every
+
+fun main() {
+    {
+        println("hi")
+    } every 10_000
+}