+++ src/test/kotlin/devtests/kSock.kt
@@ -1,0 +1,9 @@
+package devtests
+
+import klib.objects.kSock
+
+fun main() {
+    kSock.open {
+        println(it.readText().trim())
+    }
+}